Add socket option based peeloff implementation (only for APPLE currently) - build with HAVE_SCTP_PEELOFF_SOCKOPT compile option
Tested with peel_server and peel_client:  byte counts seem correct,
but checksums are wrong.

WARNING: sctp_peeloff() crashes FreeBSD 5.4, so bugs still exist
for peeloff.

+++ KERN/netinet/sctp_var.h
@@ -236,13 +236,13 @@ int sctp_usrreq __P((struct socket *, int, struct mbuf *, struct mbuf *,
 #else
 #if defined(HAVE_SCTP_SORECEIVE)
 #define sctp_sbfree(stcb, sb, m) { \
-        if((sb)->sb_cc >= (m)->m_len) { \
+        if((sb)->sb_cc >= (uint32_t)(m)->m_len) { \
   	   (sb)->sb_cc -= (m)->m_len; \
         } else { \
            (sb)->sb_cc = 0; \
         } \
         if(stcb) {\
-          if((stcb)->asoc.sb_cc >= (m)->m_len) {\
+          if((stcb)->asoc.sb_cc >= (uint32_t)(m)->m_len) {\
              (stcb)->asoc.sb_cc -= (m)->m_len; \
           } else  {\
              (stcb)->asoc.sb_cc = 0; \
@@ -251,7 +251,7 @@ int sctp_usrreq __P((struct socket *, int, struct mbuf *, struct mbuf *,
         if((sb)->sb_mbcnt >= MSIZE) { \
            (sb)->sb_mbcnt -= MSIZE; \
 	    if ((m)->m_flags & M_EXT) { \
-		if((sb)->sb_mbcnt >= (m)->m_ext.ext_size) { \
+		if((sb)->sb_mbcnt >= (uint32_t)(m)->m_ext.ext_size) { \
 		   (sb)->sb_mbcnt -= (m)->m_ext.ext_size; \
                 } else  { \
 		   (sb)->sb_mbcnt = 0; \
